Move FixtureType definitions to Base, and add FromJSON definitions.
