+++ src/lib/Test/Smoke/Types/Fixtures.hs
@@ -1,16 +1,30 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE ScopedTypeVariables #-}
 
 module Test.Smoke.Types.Fixtures where
 
 import Data.Aeson
-import Data.Aeson.Types (Parser, typeMismatch)
-import Data.Text (Text)
-import qualified Data.Text as Text
+import Data.Aeson.Types (typeMismatch)
 import Data.Vector (Vector)
 import qualified Data.Vector as Vector
+import Test.Smoke.Paths
 import Test.Smoke.Types.Base
+
+data Contents a
+  = Inline a
+  | FileLocation (RelativePath File)
+
+instance FromJSON a => FromJSON (Contents a) where
+  parseJSON s@(String _) =
+    Inline <$> parseJSON s
+  parseJSON (Object v) = do
+    maybeContents <- v .:? "contents"
+    maybeFile <- v .:? "file"
+    case (maybeContents, maybeFile) of
+      (Just _, Just _) -> fail "Expected \"contents\" or a \"file\", not both."
+      (Just contents, Nothing) -> Inline <$> parseJSON contents
+      (Nothing, Just file) -> return $ FileLocation file
+      (Nothing, Nothing) -> fail "Expected \"contents\" or a \"file\"."
+  parseJSON invalid = typeMismatch "contents" invalid
 
 data Fixture a
   = Fixture (Contents a) (Maybe Command)
@@ -21,46 +35,13 @@ newtype Fixtures a
 noFixtures :: Fixtures a
 noFixtures = Fixtures Vector.empty
 
-class FixtureType a where
-  fixtureName :: FixtureName
-  serializeFixture :: a -> Text
-  deserializeFixture :: Text -> a
-
-instance FixtureType Text where
-  fixtureName = "text"
-  serializeFixture = id
-  deserializeFixture = id
-
-instance FixtureType Status where
-  fixtureName = "exit-status"
-  serializeFixture = Text.pack . show . unStatus
-  deserializeFixture = Status . read . Text.unpack
-
-instance FixtureType StdIn where
-  fixtureName = "stdin"
-  serializeFixture = unStdIn
-  deserializeFixture = StdIn
-
-instance FixtureType StdOut where
-  fixtureName = "stdout"
-  serializeFixture = unStdOut
-  deserializeFixture = StdOut
-
-instance FixtureType StdErr where
-  fixtureName = "stderr"
-  serializeFixture = unStdErr
-  deserializeFixture = StdErr
-
-instance FixtureType a => FromJSON (Fixture a) where
+instance (FromJSON a, FixtureType a) => FromJSON (Fixture a) where
   parseJSON value@(String _) =
-    Fixture <$> parseFixtureTypeContents value <*> pure Nothing
+    Fixture <$> parseJSON value <*> pure Nothing
   parseJSON value@(Object v) =
-    Fixture <$> parseFixtureTypeContents value <*> v .:? "filter"
+    Fixture <$> parseJSON value <*> v .:? "filter"
   parseJSON invalid = typeMismatch "fixture" invalid
 
-instance FixtureType a => FromJSON (Fixtures a) where
+instance (FromJSON a, FixtureType a) => FromJSON (Fixtures a) where
   parseJSON (Array v) = Fixtures <$> mapM parseJSON v
-  parseJSON v = Fixtures . return <$> (parseJSON v :: Parser (Fixture a))
-
-parseFixtureTypeContents :: FixtureType a => Value -> Parser (Contents a)
-parseFixtureTypeContents = parseContents deserializeFixture
+  parseJSON v = Fixtures . Vector.singleton <$> parseJSON v
